Integrate SizeEstimator capability into DataType
This is a preparation step to add memory accounting for `Symbol`
instances.

`Symbol` instances can grow large due to large function trees,
or due to `ObjectType` instances with many keys. Not accounting for them
can cause memory issues.

But to add accounting for `Literal` we'll need to account the values.
Adding the capability to `DataType` will avoid having to create ad-hoc
instances of `SizeEstimator`

+++ server/src/test/java/io/crate/execution/engine/sort/SortingLimitAndOffsetProjectorTest.java
@@ -135,7 +135,7 @@ public class SortingLimitAndOffsetProjectorTest extends ESTestCase {
         );
         RowCellsAccountingWithEstimators rowAccounting =
             new RowCellsAccountingWithEstimators(
-                List.of(DataTypes.LONG, DataTypes.BOOLEAN),
+                List.of(DataTypes.INTEGER, DataTypes.BOOLEAN),
                 ConcurrentRamAccounting.forCircuitBreaker("testContext", circuitBreaker),
                 0);
